requirements.txt
autogen-agentchat>=0.4,<1.0
autogen-ext[openai]>=0.4,<1.0
ruff>=0.4.0
python-dotenv>=1.0.0
